requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tavily-python>=0.1.8
arxiv>=2.0.0
scholarly>=1.7.11
wikipedia>=1.4.0
pymed>=0.8.0
streamlit>=1.32.0 